added logger

+++ api/src/main/java/org/ednovo/gooru/domain/service/ScollectionServiceImpl.java
@@ -678,10 +678,52 @@ public class ScollectionServiceImpl extends BaseServiceImpl implements Scollecti
 
 	@Override
 	public void deleteCollection(String collectionId, User user) {
+		LOGGER.info("deleted collection staretd.........");
 		final Collection collection = this.getCollectionByGooruOid(collectionId, null);
-		LOGGER.info("deleted collection  kkk .........");
+		LOGGER.info("deleted collection 1.........");
+		rejectIfNull(collection, GL0056, _COLLECTION);
+		if (this.getOperationAuthorizer().hasUnrestrictedContentAccess(collectionId, user)) {
+			LOGGER.info("deleted collection 2.........");
+			final List<CollectionItem> collectionItems = this.getCollectionRepository().getCollectionItemByAssociation(collectionId, null, null);
+			List<CollectionItem> parentAssociations = this.getCollectionRepository().getCollectionItemByParentId(collectionId, null, null);
+			if (parentAssociations != null && parentAssociations.size() > 0) {
+				collectionItems.addAll(parentAssociations);
+			}
+			LOGGER.info("deleted collection 3.........");
+			try {
+				this.getCollectionEventLog().getEventLogs(collection.getCollectionItem(), user, collection.getCollectionType());
+			} catch (Exception e) {
+				LOGGER.error(e.getMessage());
+			}
+			LOGGER.info("deleted collection 4.........");
+			for (CollectionItem item : collectionItems) {
+				this.deleteCollectionItem(item.getCollectionItemId(), user,false);
+				if (item.getAssociatedUser() != null && !item.getAssociatedUser().getPartyUid().equals(user.getPartyUid())) {
+					getAsyncExecutor().deleteFromCache(V2_ORGANIZE_DATA + item.getAssociatedUser().getPartyUid() + "*");
+				}
+			}
+			LOGGER.info("deleted collection 5.........");
+			if (collection != null && collection.getUser() != null && collection.getSharing().equalsIgnoreCase(PUBLIC) && !collection.getCollectionType().equalsIgnoreCase(ResourceType.Type.PATHWAY.getType())) {
+				UserSummary userSummary = this.getUserRepository().getSummaryByUid(collection.getUser().getPartyUid());
+				if (userSummary != null && userSummary.getCollections() != null) {
+					userSummary.setCollections(userSummary.getCollections() <= 0 ? 0 : (userSummary.getCollections() - 1));
+					this.getUserRepository().save(userSummary);
+				}
+			}
+			/*LOGGER.info("deleted collection 6.........");
+			for (CollectionItem item : collectionItems) {
+				Collection parentCollection = item.getCollection();
+				if (parentCollection.getCollectionType().equals(FOLDER)) {
+					updateFolderSharing(parentCollection.getGooruOid());
+				}
+			} */
+			LOGGER.info("deleted collection 7.........");
+			
 			this.getCollectionRepository().remove(collection);
-		
+			LOGGER.info("deleted collection 8.........");
+		} else {
+			throw new UnauthorizedException(generateErrorMessage("GL0010"));
+		}
 
 		getAsyncExecutor().deleteFromCache(V2_ORGANIZE_DATA + user.getPartyUid() + "*");
 		getAsyncExecutor().deleteFromCache("v2-class-data-*");
